+++ package.json
@@ -5,8 +5,8 @@
     "main": "index.ts",
     "type": "module",
     "scripts": {
-        "start": "node ./index.ts",
-        "start:dev": "nodemon ./index.ts",
+        "start": "ts-node index.ts",
+        "start:dev": "nodemon index.ts",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Student1",
