+++ src/model.ts
@@ -1,0 +1,11 @@
+export enum cmdList {
+  LEFT = 'mouse_left',
+  RIGHT = 'mouse_right',
+  UP = 'mouse_up',
+  DOWN = 'mouse_down',
+  POSITION = 'mouse_position',
+  SQUARE = 'draw_square',
+  RECTANGLE = 'draw_rectangle',
+  CIRCLE = 'draw_circle',
+  SCREEN = 'prnt_scrn'
+};
